64-bit updates

+++ Networking/NetworkingSource.m
@@ -38,7 +38,7 @@
 		NSMutableArray *addresses = [NSMutableArray arrayWithCapacity:1];
 		NSString *testaddr;
 		struct ifaddrs *addr0, *addr;
-		int family, result;
+		NSInteger family, result;
 		char ipaddr[NI_MAXHOST];
 		if (getifaddrs(&addr0) == -1) {
 			return nil;
